Create gaussian images described in the paper
The gaussian images are created as described in
the paper "The Laplacian Pyramid as a Compact Image Code".
The presented formula in the paper to calculate the dimension fo an image at a
specific level was changed to handle indexes starting with 0.
It was also changed in the manner of a fix, because the original formula seems wrong.